requirements.txt
# ============================================================================
# DEPENDENCIAS - Sistema RRHH v2.0 + Supabase
# ============================================================================

# Análisis de datos
pandas>=1.3.0
numpy>=1.21.0

# Visualización
plotly>=5.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Dashboard (si usas Dash)
dash>=2.14.0
dash-bootstrap-components>=1.5.0

# Supabase para base de datos online
supabase>=2.3.0

# Google Sheets (opcional)
gspread>=5.0.0
oauth2client>=4.1.3
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0

# Utilidades
python-dateutil>=2.8.2
pytz>=2021.3

# Para desarrollo y testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
pylint>=2.12.0
ipython>=8.0.0
ipykernel>=6.0.0

# Monitoreo de recursos
psutil>=5.9.0

# HTTP y APIs
requests>=2.28.0

# Gestión de entornos
python-dotenv>=0.19.0
httpx==0.27.2